requirements.txt
# Core data processing
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Deep learning
tensorflow>=2.8.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Utilities
pyyaml>=6.0
tqdm>=4.62.0
colorlog>=6.6.0

# Optional: For Jupyter notebooks
jupyter>=1.0.0
ipywidgets>=7.7.0